docker/scraper/requirements.txt
protobuf==3.20.3
playwright==1.41.2
kafka-python==2.0.2
scrapy==2.11.0